Refactory postpone method
