+++ modules.js
@@ -216,3 +216,134 @@ export function dateDiff(originalDate, futureDate) {
   const _MS_PER_DAY = 1000 * 60 * 60 * 24;
   return Math.floor((futureDate - originalDate) / _MS_PER_DAY);
 }
+
+export function postponeByRules(json) {
+  let putJson = {};
+
+  // 1. checklist overdue -> due + postponable (consider action days): if still overdue => today
+  //    checklist due today -> due + postponable (consider action days)
+  const actionDays = json.customFields.["Action days"] ? json.customFields.["Action days"] : "Any day";
+  const daysPostponable = setDaysPostponable(json.customFields.["Days postponable"], json.customFields.Priority);
+  const recurring = json.customFields.Recurring ? parseInt(json.customFields.Recurring) : 0;
+  const today = new Date();
+  const checkListHours = 21;
+  let earlierDate;
+  let laterDate;
+
+  today.setUTCHours(0,0,0,0);
+
+  putJson.checkListItems = [];
+
+  for (let i = 0; i < json.checkListItems.length; i ++) {
+    let dueDate = new Date(json.checkListItems[i].due);
+    const tomorrow = addDays(today, 1, actionDays, today);
+
+    if(dueDate < tomorrow) {
+      dueDate = addDays(dueDate, daysPostponable, actionDays, today);
+      dueDate.setUTCHours(checkListHours);
+      json.checkListItems[i].due = dueDate;
+
+      putJson.checkListItems.push({
+        id : json.checkListItems[i].id,
+        params : {
+          due : JSON.parse(JSON.stringify(dueDate))
+        }
+      })
+    }
+
+    earlierDate = (dueDate < earlierDate || !earlierDate) ? dueDate : earlierDate;
+    laterDate = (dueDate > laterDate || !laterDate) ? dueDate : laterDate;
+  }
+
+  // 2. next action -> same date as earlier checklist item
+  putJson.customFields = [];
+  let nextAction = new Date(json.customFields.["Next action"]);
+
+  if (json.customFields.["Next action"]) {
+    nextAction.setUTCFullYear(earlierDate.getUTCFullYear());
+    nextAction.setUTCMonth(earlierDate.getUTCMonth());
+    nextAction.setUTCDate(earlierDate.getUTCDate());  
+  } else {
+    nextAction = earlierDate;
+  }
+
+  if (JSON.stringify(nextAction) != JSON.stringify(json.customFields.["Next action"])) {
+    json.customFields.["Next action"] = nextAction;
+
+    putJson.customFields.push({
+      idCustomField : json.customFields.["idCustomFieldNext action"],
+      body : {
+        value : {
+          date : JSON.parse(JSON.stringify(json.customFields.["Next action"]))
+        }
+      }
+    });
+  
+  };
+
+  // 3. due date -> if lower than later date => original date + postponable weeks -> if greater than today + recurring => later date
+  let due = new Date(json.due);
+
+  if (due < laterDate || !json.due) {
+
+    if (!json.due) {
+      due = laterDate;
+    } else {
+      due.setUTCDate(due.getUTCDate() + daysPostponable * 7);
+    }
+
+    let nextRecurrent = new Date(today);
+    nextRecurrent.setUTCDate(nextRecurrent.getUTCDate() + recurring);
+
+    if (recurring > 0 && due > nextRecurrent) {
+      due.setUTCFullYear(laterDate.getUTCFullYear());
+      due.setUTCMonth(laterDate.getUTCMonth());
+      due.setUTCDate(laterDate.getUTCDate());
+    }
+
+  }
+
+  if (JSON.stringify(due) != JSON.stringify(json.due)) {
+    json.due = due;
+
+    putJson.main = {
+      params : {
+        due : JSON.parse(JSON.stringify(json.due))
+      }
+    };
+  
+  };
+
+  return putJson;
+}
+
+function setDaysPostponable(daysPostponable, priority) {
+  let days;
+
+  if(daysPostponable) {
+    days = parseInt(daysPostponable);
+  } else {
+
+    switch(priority) {
+
+      case "Urgent":
+        days = 1;
+        break;
+
+      case "Important":
+        days = 2;
+        break;
+
+      case "Medium":
+        days = 3;
+        break;
+
+      default:
+        days = 4;
+        break;
+
+    }
+  }
+
+  return days;
+}
